account-service/src/test/java/ru/ashabalin/account/service/controllers/AccountControllerTest.java
package ru.ashabalin.account.service.controllers;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.web.servlet.MockMvc;
import ru.ashabalin.account.service.services.AccountService;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@AutoConfigureMockMvc
@ContextConfiguration(classes = {AccountController.class})
@WebMvcTest
public class AccountControllerTest {

    @MockBean
    private AccountService accountService;

    @Autowired
    private MockMvc mockMvc;

    @Test
    public void getAmountTest() throws Exception {
        when(accountService.getAmount(1)).thenReturn(10L);
        String contentAsString = mockMvc.perform(get("/api/v1/account/1"))
                .andExpect(status().isOk())
                .andReturn()
                .getResponse()
                .getContentAsString();

        assertThat(contentAsString).isEqualTo("10");
    }

    @Test
    public void addAmountTest() throws Exception {
        mockMvc.perform(post("/api/v1/account/1?value=10"))
                .andExpect(status().isOk());

        verify(accountService).addAmount(1, 10L);
    }
}
